chore: move all changes to CompositionTarget
