+++ src/Uno.UI.Composition/Composition/ICompositionTarget.cs
@@ -1,6 +1,8 @@
+#nullable enable
 using System;
 using System.Collections.Generic;
 using Windows.ApplicationModel.VoiceCommands;
+using Windows.Foundation;
 using Microsoft.UI.Composition.Interactions;
 
 namespace Uno.UI.Composition;
@@ -11,7 +13,7 @@ internal interface ICompositionTarget
 
 	double RasterizationScale { get; }
 
-	event EventHandler RasterizationScaleChanged;
+	event EventHandler? RasterizationScaleChanged;
 
 #if __SKIA__
 	void RequestNewFrame();
