--- src/Uno.UI.Dispatching/DispatchingFeatureConfiguration.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.ComponentModel;
-
-namespace Uno;
-
-public static class DispatchingFeatureConfiguration
-{
-	internal static class DispatcherQueue
-	{
-		/// <summary>
-		/// This property is used to throttle the number of frames dispatched
-		/// to the UI thread under webassembly.
-		/// </summary>
-		public static double WebAssemblyFrameRate { get; set; } = 60.0;
-	}
-}
