+++ src/Uno.UI/FeatureConfiguration.cs
@@ -93,24 +93,11 @@ namespace Uno.UI
 
 		public static class CompositionTarget
 		{
-			private static float _frameRate = 60;
-
 			/// <summary>
 			/// Suggested frame rate for <see cref="Microsoft.UI.Xaml.Media.CompositionTarget.Rendering"/> event.
 			/// This property is used by desktop skia renderers.
 			/// </summary>
-			public static float FrameRate
-			{
-				get => _frameRate;
-				set
-				{
-					_frameRate = value;
-
-					// Use this because we do not depend on the wasm dispatching binaries
-					// at this layer (we use the reference binaries).
-					DispatchingFeatureConfiguration.DispatcherQueue.WebAssemblyFrameRate = _frameRate;
-				}
-			}
+			public static float FrameRate { get; set; } = 60;
 		}
 
 		public static class ContentPresenter
