+++ src/Uno.UI/UI/Xaml/Media/CompositionTarget.cs
@@ -16,6 +16,19 @@ public partial class CompositionTarget : ICompositionTarget
 	internal CompositionTarget(ContentRoot contentRoot)
 	{
 		ContentRoot = contentRoot;
+#if __SKIA__
+		RenderingActiveChanged += ((ICompositionTarget)this).RequestNewFrame;
+		// TODO: is HostWindow ever null?
+		if (ContentRoot.XamlRoot?.HostWindow is { } hostWindow)
+		{
+			void OnHostWindowClosed(object sender, WindowEventArgs args)
+			{
+				RenderingActiveChanged -= ((ICompositionTarget)this).RequestNewFrame;
+				hostWindow.Closed -= OnHostWindowClosed;
+			}
+			hostWindow.Closed += OnHostWindowClosed;
+		}
+#endif
 	}
 
 	event EventHandler ICompositionTarget.RasterizationScaleChanged
